requirements.txt
alabaster==0.7.12
argh==0.26.2
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
idna==2.8
imagesize==1.1.0
Jinja2==2.10
lambda-git==0.1.1
livereload==2.6.0
MarkupSafe==1.1.1
packaging==19.0
pathtools==0.1.2
port-for==0.3.1
Pygments==2.3.1
pyparsing==2.3.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinx-argparse==0.2.5
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-programoutput==0.13
sphinxcontrib-websupport==1.1.0
tornado==6.0.1
urllib3==1.24.1
watchdog==0.9.0
